requirements.txt
python-dotenv
PyPDF2
fastapi
uvicorn[standard]
sqlalchemy
pydantic-ai-slim[duckduckgo]
duckduckgo-search>=5.0.0
supabase
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic[email]
pydantic
pydantic_ai
langchain
langgraph
langchain-groq
psycopg2-binary
PyJWT
langchain-google-genai
urllib3
langchain-community
python-jose[cryptography]
langchain-pinecone
pinecone
google-genai
docx
neo4j
supabase
alembic
qdrant-client
supabase
python-docx
pdfplumber
feedparser
httpx
beautifulsoup4
selectolax

